t2400: Fix test failures when using grep 2.5
Replace all cases of `\s` with `[[:space:]]` as older versions of GNU
grep (and from what it seems most versions of BSD grep) do not handle
`\s`.

For the same reason all cases of `\S` are replaced with `[^[:space:]]`.
Replacing `\S` also needs to occur as `\S` is technically PCRE and not
part of ERE even though most modern versions of grep accept it as ERE.

Signed-off-by: Jacob Abel <jacobabel@nullpo.dev>