t2400: Fix test failures when using grep 2.5
Replace all cases of `\s` with `[[:blank:]]` or ` ` as older versions
of GNU grep (and from what it seems most versions of BSD grep) do not
handle `\s`.

For the same reason all cases of `\S` are replaced with `[^ ]`. It's not
an exact replacement (as it does not match tabs) but it is close enough
for this use case.

Replacing `\S` also needs to occur as `\S` is technically PCRE and not
part of ERE even though most modern versions of grep accept it as ERE.

This commit also drops `--sq` from a rev-parse call as it appears to be
a no-op.

Signed-off-by: Jacob Abel <jacobabel@nullpo.dev>